Fix feedback Ivan for current MR
