+++ src/app/admin/users/service.ts
@@ -96,7 +96,7 @@ export const useUserDelete = (config: MutationOptions = {}) => {
 export const useUserCreate = (config: MutationOptions = {}) => {
   return useMutation<any, any, any>(
     ({ login, firstName, lastName, email, langKey = 'en', authorities }) =>
-      Axios.post('/admin/users', {
+      Axios.post('/users', {
         login,
         firstName,
         lastName,
