+++ src/components/MenuAction/index.tsx
@@ -1,6 +1,72 @@
 import React, { FC } from 'react';
 
-import { MenuProps, useDisclosure, Menu, MenuButton } from '@chakra-ui/react';
+import {
+  Box,
+  MenuItemProps,
+  MenuItem,
+  Collapse,
+  Heading,
+  Flex,
+  Button,
+  MenuProps,
+  useDisclosure,
+  Menu,
+  MenuButton,
+} from '@chakra-ui/react';
+
+interface IMenuItemProps extends MenuItemProps {
+  propsActionHeader: any;
+  propsConfirmButton: any;
+  propsCancelButton: any;
+  confirmationText: string;
+  text;
+  actionCallBack: any;
+  menuAction: {
+    state: {
+      isOpen: boolean;
+      onToggle: any;
+      onClose: any;
+    };
+    action: any;
+  };
+}
+
+export const MenuActionItem: FC<IMenuItemProps> = ({
+  propsActionHeader,
+  propsConfirmButton,
+  propsCancelButton,
+  confirmationText,
+  actionCallBack,
+  menuAction,
+  text,
+  ...rest
+}) => {
+  const { isOpen, onToggle } = menuAction.state || {
+    isOpen: null,
+    onToggle: null,
+    onClose: null,
+  };
+  return (
+    <>
+      <MenuItem onClick={onToggle} {...rest}>
+        {text}
+      </MenuItem>
+      <Collapse in={isOpen} animateOpacity>
+        <Box bg="gray.50" py="2" px="4">
+          <Heading {...propsActionHeader}>{confirmationText}</Heading>
+          <Flex>
+            <Button {...propsCancelButton} onClick={onToggle}>
+              {propsCancelButton.text}
+            </Button>
+            <Button {...propsConfirmButton} onClick={actionCallBack}>
+              {propsConfirmButton.text}
+            </Button>
+          </Flex>
+        </Box>
+      </Collapse>
+    </>
+  );
+};
 
 export interface MenuActionProps extends MenuProps {}
 
@@ -24,7 +90,7 @@ export const useMenuAction = (menuActions: any) => {
     });
   };
 
-  return [callBackConfirmButton, onCloseMenu, onToggleMenu, isOpenMenu];
+  return { callBackConfirmButton, onCloseMenu, onToggleMenu, isOpenMenu };
 };
 
 export const MenuAction: FC<any> = ({
